Create package for models
